Prepare code for label cleaning and processing
This commit changes some functions to adapt the PC landsat code to NAIP
and also to clean some of the labels. We create a main function to
eliminate houses with overlapping buffers.

We need to make still changes in the PC code to take the image and use
rasterio to crop it to the buffer bounds. Test code in the MS PC Hub,
but nothing in here, yet.