Many changes to make things work
Changes in prompting strategy to make binary prediction rather than
generation of text. The prompt is now optional, so is the reponse label
at the end. Now, we also add new configs for training just to feed a
Config class. The dataset is revamped and the collator function is
killed!

gpt-2 is now trained with a simple file!