Changes in ViT training -- sweeps!
Do hyper-parameter selection on ViT to improve the model behavior. Loss
seems to be better now, but we need to test with full data in a bigger
computer. Hence, this commit!

CustomTrainer is new defult and is taking weights from the data to pass
them into the CE loss function.